Pin clang-format to v18 and switch to C++20
Match the host build to the device firmware's gnu++20 standard. Unit
tests then exercise the exact language rules the firmware compiles
under, and unblocks C++20 standard library usage such as
std::string::ends_with.

Pin the local formatting toolchain to 18.1.8 to match CI (ubuntu-24.04).
clang-format 19+ switched ColumnLimit from bytes to display columns,
which causes formatting drift for sources containing multi-byte
punctuation near the 120-column boundary.

Move wait_for_air_time_ to an anonymous namespace as a free function,
and replace magic constants in soft_phy_air_time_us with explicit
line-rate and time-unit definitions.

+++ components/home_io_control/pairing_engine.cpp
@@ -431,7 +431,7 @@ bool PairingEngine::discover_and_pair() {
 
   // Phase 2: Key exchange — retry up to the configured number of times.
   bool key_exchanged = false;
-  for (int ke_attempt = 0; ke_attempt < tuning_->pairing_key_exchange_retries; ke_attempt++) {
+  for (uint8_t ke_attempt = 0; ke_attempt < tuning_->pairing_key_exchange_retries; ke_attempt++) {
     if (ke_attempt > 0) {
       ESP_LOGI(TAG, "Retrying key exchange (attempt %d/%u)...", ke_attempt + 1, tuning_->pairing_key_exchange_retries);
       App.feed_wdt();
